backend/requirements.txt
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
elasticsearch==8.2.2
elastic-transport==8.1.2